Fix a coverity and a build warning

+++ src/strings.c
@@ -1265,7 +1265,7 @@ string_from_blob(blob_T *blob, long *start_idx)
 {
     garray_T	str_ga;
     long	blen;
-    long	idx;
+    int		idx;
 
     ga_init2(&str_ga, sizeof(char), 80);
